requirements.txt
Django==2.2.8
django-environ==0.4.5
djangorestframework==3.11.0
celery==4.4.0
mercadopago==0.3.5
psycopg2>=2.7,<2.8 --no-binary psycopg2
requests==2.22.0
uuid==1.30
django-ckeditor==5.8.0
pyjwt[crypto]==1.7.1
